lib/vfscore: libc-style `ioctl`: Typecast arguments
Avoids a compiler warning because of missing typecasts of
arguments hand-over to the actual implementing function
`uk_syscall_e_ioctl()`.

+++ lib/vfscore/main.c
@@ -782,7 +782,7 @@ int ioctl(int fd, unsigned long int request, ...)
 	arg = va_arg(ap, void*);
 	va_end(ap);
 
-	return uk_syscall_e_ioctl(fd, request, arg);
+	return uk_syscall_e_ioctl((long) fd, (long) request, (long) arg);
 }
 #endif
